+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Vendor
 
 * Update SPC version to 2.3.0
+* Update all PHP vendor
 
 ### Fixes